added check

+++ game_trends/trends_app/models.py
@@ -82,12 +82,12 @@ def checkTopGames():
 		game.save()
 	
 	low_ranked_games = Game.objects.all().annotate(total_rank=F('play_rank')+F('growth_rank')+F('h_rank')+F('h_growth_rank')).order_by('-total_rank')
-
-	for i in range(len(low_ranked_games)-max_size):
-		if low_ranked_games[i].fav_users.all().count()>0:
-			continue
-		else:
-			low_ranked_games[i].delete()
+	if len(low_ranked_games)>max_size:
+		for i in range(len(low_ranked_games)-max_size):
+			if low_ranked_games[i].fav_users.all().count()>0:
+				continue
+			else:
+				low_ranked_games[i].delete()
 	return checkTopGames()
 	#return
